Fix weight_decay (sign convention) in CMAES
Due to the sign-convention of the L2 norm -> L2 Decay in `compute_weight_decay`, line 25, the `l2_decay` needs to be **added to the original `reward_table` values**, not to the negative ones. The conversion from minimizer to maximize happens subsequently. Otherwise, the parameters will blow up. Btw: huge fan here ...